copier copy port with testing part 1
Adds tests and removes object interface dependency while embedding the
full CopyObjectInput so we can set any flags/settings that we want in a
copy.

- Add singlePartcopyObject and tests
- Fix import and update deps for merge
- Flesh out tests for getContentLength & objectInfo
- Use CopyOjectWithContext and embed CopyObjectInput

We use the WithContextMethods so we can cancel in flight items if we
exit early, e.g. on a signal. Also we want to be able to set any s3
options on copy so embed the full s3.CopyObjectInput struct.